requirements.txt
discord.py==2.0.1
google-cloud-firestore==2.4.0
google-api-python-client==2.52.0
python-dotenv==0.19.2
selenium==4.1.0
